re-layer logstream for web and function apps
