+++ client-react/src/utils/scenario-checker/function-app.environment.ts
@@ -60,6 +60,13 @@ export class FunctionAppEnvironment extends Environment {
         return { status: 'enabled' };
       },
     };
+
+    this.scenarioChecks[ScenarioIds.showFunctionAppLogs] = {
+      id: ScenarioIds.showFunctionAppLogs,
+      runCheck: () => {
+        return { status: 'enabled' };
+      },
+    };
   }
 
   public isCurrentEnvironment(input?: ScenarioCheckInput): boolean {
