+++ client-react/src/utils/scenario-checker/scenario-ids.ts
@@ -109,4 +109,5 @@ export class ScenarioIds {
   public static readonly runtimeScaleMonitoringSupported = 'runtimeScaleMonitoringSupported';
   public static readonly xenonAppRuntimeStack = 'xenonAppRuntimeStack';
   public static readonly showAppInsightsLogs = 'showAppInsightsLogs';
+  public static readonly showFunctionAppLogs = 'showFunctionAppLogs';
 }
